Adapted CLI to use the new MemoryInterface.

+++ src/dap_access.rs
@@ -5,5 +5,5 @@ pub trait DAPAccess {
     fn read_register(&mut self, port: u16, addr: u32) -> Result<u32, Self::Error>;
 
     /// Writes a value to the DAP register on the specified port and address
-    fn write_register(&mut self, port: u16, addr: u16, value: u32) -> Result<(), Self::Error>;
+    fn write_register(&mut self, port: u16, addr: u32, value: u32) -> Result<(), Self::Error>;
 }
